+++ lib/src/builders/base/base_notifier_builder.dart
@@ -1,22 +1,52 @@
+// ignore_for_file: invalid_use_of_protected_member
+
 import 'package:flutter/widgets.dart'
-    show StatefulWidget, Widget, BuildContext, VoidCallback, State, SizedBox;
+    show
+        BuildContext,
+        SizedBox,
+        State,
+        StatefulWidget,
+        VoidCallback,
+        Widget,
+        WidgetsBinding;
 import 'package:syncx/src/notifier/base/base_notifier.dart' show BaseNotifier;
 import 'package:syncx/src/utils/extensions/build_context_extensions.dart';
 
-/// A base class for building widgets that listen to a [BaseNotifier] and rebuild
-/// when its state changes. Intended to be extended by more specific notifier builders.
+part '../sync/notifier_builder.dart';
+part '../sync/notifier_consumer.dart';
+part '../sync/notifier_listener.dart';
+
+/// A base widget for listening to a [BaseNotifier] and rebuilding when its state changes.
+///
+/// [BaseNotifierBuilder] is intended to be extended by more specific notifier builders, such as [NotifierBuilder] or [AsyncNotifierBuilder].
+/// It manages the lifecycle of the notifier, listens for state changes, and provides hooks for rebuilding and initialization.
 ///
 /// [N] is the type of [BaseNotifier] and [S] is the type of state managed by the notifier.
-abstract class BaseNotifierBuilder<N extends BaseNotifier<S>, S>
-    extends StatefulWidget {
+abstract class _BaseNotifierBuilder<N extends BaseNotifier<S>,
+    S extends Object?> extends StatefulWidget {
   /// Creates a [BaseNotifierBuilder].
   ///
-  /// [notifier] is the instance of [BaseNotifier] to listen to. If null, it will be obtained from the nearest [Provider].
-  ///
-  /// [child] is an optional static child widget.
-  ///
+  /// [notifier] provides the [BaseNotifier] instance to listen to. If null, the nearest [Provider] is used.
+  /// [builder] is called to build the widget tree based on the current state.
   /// [onInit] is an optional callback invoked with the notifier when the widget is initialized.
+  /// [child] is an optional static child widget.
+  /// [key] is the widget key.
   ///
+  const _BaseNotifierBuilder({
+    this.notifier,
+    this.builder,
+    this.onInit,
+    this.child,
+    super.key,
+  });
+
+  /// Provides the [BaseNotifier] instance to listen to. If null, the nearest [Provider] is used.
+  final N Function()? notifier;
+
+  /// An optional static child widget to display if [builder] is not provided.
+  final Widget? child;
+
+  /// An optional callback invoked with the notifier when the widget is initialized.
   /// Note: Do not manually call the notifier's `onInit()` method here. The notifier itself will automatically
   /// trigger its own `onInit()` when it is created, so calling it again would result in duplicate invocations.
   ///
@@ -28,30 +58,10 @@ abstract class BaseNotifierBuilder<N extends BaseNotifier<S>, S>
   ///   builder: (context, state) => Text('State: $state'),
   /// )
   /// ```
-  ///
-  /// [builder] is a function that builds the widget tree based on the current state.
-  const BaseNotifierBuilder({
-    this.notifier,
-    this.builder,
-    this.onInit,
-    this.child,
-    super.key,
-  });
-
-  /// The [BaseNotifier] instance to listen to. If null, it will be obtained from the nearest [Provider].
-  final N? notifier;
-
-  /// An optional static child widget.
-  final Widget? child;
-
-  /// An optional callback invoked with the notifier when the widget is initialized.
-  ///
-  /// Note: Do not manually call the notifier's `onInit()` method here. The notifier itself will automatically
-  /// trigger its own `onInit()` when it is created, so calling it again would result in duplicate invocations.
   final void Function(N notifier)? onInit;
 
   /// A function that builds the widget tree based on the current state.
-  final Widget Function(BuildContext context, S state)? builder;
+  final Widget Function(S state)? builder;
 
   /// Called when the notifier's state changes.
   ///
@@ -63,11 +73,13 @@ abstract class BaseNotifierBuilder<N extends BaseNotifier<S>, S>
   State<StatefulWidget> createState() => _BaseNotifierBuilderState<N, S>();
 }
 
-/// The state class for [BaseNotifierBuilder]. Handles listening to the notifier and rebuilding when needed.
+/// State class for [BaseNotifierBuilder].
+///
+/// Handles listening to the notifier, managing state, and triggering rebuilds when needed.
 class _BaseNotifierBuilderState<N extends BaseNotifier<S>, S>
-    extends State<BaseNotifierBuilder<N, S>> {
+    extends State<_BaseNotifierBuilder<N, S>> {
   /// The notifier instance being listened to. Initialized from the widget or the nearest Provider.
-  late N notifier = widget.notifier ?? context.get<N>();
+  late N notifier = widget.notifier?.call() ?? context.get<N>();
 
   /// The previous state, used to compare with the current state on changes.
   late S _previous = notifier.state;
@@ -75,29 +87,32 @@ class _BaseNotifierBuilderState<N extends BaseNotifier<S>, S>
   @override
   void initState() {
     super.initState();
-    notifier.addListener(_whenStateChanged);
-    widget.onInit?.call(notifier);
+    _addListener();
+    _onInit();
   }
 
   @override
-  void didUpdateWidget(covariant BaseNotifierBuilder<N, S> oldWidget) {
+  void didUpdateWidget(covariant _BaseNotifierBuilder<N, S> oldWidget) {
     super.didUpdateWidget(oldWidget);
     // If the notifier instance changes, update the listener.
-    if (widget.notifier == null || widget.notifier == notifier) return;
-    notifier = widget.notifier!;
+    final newNotifier = widget.notifier?.call();
+    if (newNotifier == null || newNotifier == notifier) return;
+    _removeListener();
+    notifier = newNotifier;
+    _addListener();
   }
 
   @override
   Widget build(BuildContext context) {
     // Build using the builder function if provided, otherwise use the child or an empty box.
-    return widget.builder?.call(context, _previous) ??
+    return widget.builder?.call(_previous) ??
         widget.child ??
         const SizedBox.shrink();
   }
 
   @override
   void dispose() {
-    notifier.removeListener(_whenStateChanged);
+    _removeListener();
     super.dispose();
   }
 
@@ -107,12 +122,22 @@ class _BaseNotifierBuilderState<N extends BaseNotifier<S>, S>
     super.setState(fn);
   }
 
+  void _onInit() {
+    if (widget.onInit == null) return;
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      widget.onInit!.call(notifier);
+    });
+  }
+
   /// Internal callback when the notifier's state changes.
   void _whenStateChanged() {
-    widget.whenStateChanged(_previous, notifier.state, _reBuild);
+    if (mounted) widget.whenStateChanged(_previous, notifier.state, _reBuild);
     _previous = notifier.state;
   }
 
   /// Triggers a rebuild of the widget.
   void _reBuild() => setState(() {});
+
+  void _addListener() => notifier.addListener(_whenStateChanged);
+  void _removeListener() => notifier.removeListener(_whenStateChanged);
 }