+++ lib/src/register/notifier_register.dart
@@ -7,8 +7,8 @@ import 'package:syncx/src/notifier/base/base_notifier.dart';
 ///
 /// [NotifierRegister] is a convenience widget for providing a notifier to its descendants.
 ///
-/// [N] is the type of [BaseNotifier] and [S] is the type of state managed by the notifier.
-class NotifierRegister<N extends BaseNotifier<S>, S extends Object?>
+/// [N] is the type of [BaseNotifier].
+class NotifierRegister<N extends BaseNotifier<Object?>>
     extends StatelessWidget {
   /// Creates a [NotifierRegister].
   ///
@@ -40,6 +40,7 @@ class NotifierRegister<N extends BaseNotifier<S>, S extends Object?>
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider(
+      key: key,
       lazy: lazy,
       create: create,
       builder: builder,