+++ lib/src/utils/models/error_state.dart
@@ -11,4 +11,21 @@ class ErrorState {
 
   /// An optional stack trace associated with the error.
   final StackTrace? stackTrace;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    return other is ErrorState &&
+        other.error == error &&
+        other.message == message &&
+        other.stackTrace == stackTrace;
+  }
+
+  @override
+  int get hashCode => Object.hash(error, message, stackTrace);
+
+  @override
+  String toString() {
+    return 'ErrorState(error: $error, message: $message, stackTrace: $stackTrace)';
+  }
 }